+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v0.8
 
 * added `are_mutually_nonintersecting` to `geom3d.paths`
+* bump satella
 
 # v0.7